+++ frontend/src/components/dating/DatingRandomizationDesktopPage.tsx
@@ -1,0 +1,7 @@
+import DatingAppBody from "./DatingAppBody"
+
+const DatingRandomizationDesktopPage = () => {
+    return <DatingAppBody>Randomization (Desktop)</DatingAppBody>
+}
+
+export default DatingRandomizationDesktopPage
